Various fixes and improvements
Added a port listening service to allow only on instance of the
application to open at once. Moved background drawing to main lop to fix
menu. The research menu is hidden until the user presses Shift + R and
is hidden with Shift + S. Code was added to setting and checking the
pulse settings and creating the dialog